Create a LoadingNode to reduce code duplication.
