+++ libraries/architecture/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
     api(libs.appyx.core)
     api(libs.androidx.lifecycle.runtime)
     api(libs.molecule.runtime)
+    implementation(projects.libraries.designsystem)
 
     testCommonDependencies(libs)
 }
